Fixed issue on PR #7 Added icons.dart file Changed util code from middleware to a separate class Removed unused imports
