+++ lib/screens/assets/icons.dart
@@ -1,0 +1,6 @@
+const imgSvgIconAlarm = "assets/ico_alarm.svg";
+const imgSvgIconFire = "assets/ico_fire.svg";
+const imgSvgIconWatch = "assets/ico_watch.svg";
+const imgSvgLogoFlame = "assets/logo_flame.svg";
+const imgSvgIconFake = "assets/ico_fake.svg";
+const imgSvgIconPointer = "assets/ico_pointer.svg";
