+++ lib/screens/assets/images.dart
@@ -1,12 +1,6 @@
 const imgSvgFireman = "assets/fireman.svg";
-const imgSvgIconAlarm = "assets/ico_alarm.svg";
-const imgSvgIconFire = "assets/ico_fire.svg";
-const imgSvgIconWatch = "assets/ico_watch.svg";
-const imgSvgLogoFlame = "assets/logo_flame.svg";
-const imgSvgPlane = "assets/plane.svg";
 const imgSvgFireTruck = "assets/firetruck.svg";
-const imgSvgIconFake = "assets/ico_fake.svg";
-const imgSvgIconPointer = "assets/ico_pointer.svg";
+const imgSvgPlane = "assets/plane.svg";
 const imgSvgLogo = "assets/logo.svg";
 const imgLightning = "assets/lightning.png";
 const imgLogo = "assets/logo.png";
